+++ django_webix_sender/send_methods/storage/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from django_webix_sender.send_methods.storage.send_utils import send
+
+__all__ = (
+    'send'
+)